pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "bcrypt>=4.2.1",
    "flask>=3.1.0",
    "flask-cors>=5.0.0",
    "flask-session>=0.8.0",
    "flask-sqlalchemy>=3.1.1",
    "passlib>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.10.6",
    "python-dotenv>=1.0.1",
    "python-jose>=3.3.0",
    "sqlalchemy>=2.0.37",
]
